+++ src/textual/widgets/_input.py
@@ -2,7 +2,8 @@ from __future__ import annotations
 
 import re
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, ClassVar, Iterable
+from enum import IntFlag
+from typing import TYPE_CHECKING, ClassVar, Iterable, Pattern
 
 from rich.cells import cell_len, get_character_cell_size
 from rich.console import Console, ConsoleOptions, RenderableType
@@ -23,7 +24,7 @@ from ..css._error_tools import friendly_list
 from ..events import Blur, Focus, Mount
 from ..geometry import Offset, Size
 from ..message import Message
-from ..reactive import reactive, var
+from ..reactive import Reactive, reactive, var
 from ..suggester import Suggester, SuggestionReady
 from ..timer import Timer
 from ..validation import ValidationResult, Validator
@@ -43,6 +44,41 @@ _RESTRICT_TYPES = {
 InputType = Literal["integer", "number", "text"]
 
 
+class _CharFlags(IntFlag):
+    """Misc flags for a single template character definition"""
+
+    REQUIRED = 0x1
+    """Is this character required for validation?"""
+
+    SEPARATOR = 0x2
+    """Is this character a separator?"""
+
+    UPPERCASE = 0x4
+    """Char is forced to be uppercase"""
+
+    LOWERCASE = 0x8
+    """Char is forced to be lowercase"""
+
+
+_TEMPLATE_CHARACTERS = {
+    "A": (r"[A-Za-z]", _CharFlags.REQUIRED),
+    "a": (r"[A-Za-z]", None),
+    "N": (r"[A-Za-z0-9]", _CharFlags.REQUIRED),
+    "n": (r"[A-Za-z0-9]", None),
+    "X": (r"[^ ]", _CharFlags.REQUIRED),
+    "x": (r"[^ ]", None),
+    "9": (r"[0-9]", _CharFlags.REQUIRED),
+    "0": (r"[0-9]", None),
+    "D": (r"[1-9]", _CharFlags.REQUIRED),
+    "d": (r"[1-9]", None),
+    "#": (r"[0-9+\-]", None),
+    "H": (r"[A-Fa-f0-9]", _CharFlags.REQUIRED),
+    "h": (r"[A-Fa-f0-9]", None),
+    "B": (r"[0-1]", _CharFlags.REQUIRED),
+    "b": (r"[0-1]", None),
+}
+
+
 class _InputRenderable:
     """Render the input content."""
 
@@ -60,6 +96,7 @@ class _InputRenderable:
         # Add the completion with a faded style.
         value = input.value
         value_length = len(value)
+        template = input._template
         suggestion = input._suggestion
         show_suggestion = len(suggestion) > value_length and input.has_focus
         if show_suggestion:
@@ -67,6 +104,15 @@ class _InputRenderable:
                 suggestion[value_length:],
                 input.get_component_rich_style("input--suggestion"),
             )
+        elif template is not None:
+            style = input.get_component_rich_style("input--placeholder")
+            result += Text(
+                template.mask[value_length:],
+                style,
+            )
+            for index, (c, char_def) in enumerate(zip(value, template.template)):
+                if c == " ":
+                    result.stylize(style, index, index + 1)
 
         if self.cursor_visible and input.has_focus:
             if not show_suggestion and input._cursor_at_end:
@@ -88,6 +134,251 @@ class _InputRenderable:
             line_length,
         )
         yield from line
+
+
+class _Template(Validator):
+    """Template mask enforcer."""
+
+    @dataclass
+    class CharDef:
+        """Holds data for a single char of the template mask."""
+
+        pattern: Pattern[str]
+        """Compiled regular expression to check for matches."""
+
+        flags: _CharFlags = _CharFlags(0)
+        """Flags defining special behaviors"""
+
+        char: str = ""
+        """Mask character (separator or blank or placeholder)"""
+
+    def __init__(self, input: Input, template_str: str) -> None:
+        self.input = input
+        self.template: list[_Template.CharDef] = []
+        self.blank: str = " "
+        escaped = False
+        flags = _CharFlags(0)
+        template_chars: list[str] = list(template_str)
+
+        while template_chars:
+            c = template_chars.pop(0)
+            if escaped:
+                char = self.CharDef(re.compile(re.escape(c)), _CharFlags.SEPARATOR, c)
+                escaped = False
+            else:
+                if c == "\\":
+                    escaped = True
+                    continue
+                elif c == ";":
+                    break
+
+                new_flags = {
+                    ">": _CharFlags.UPPERCASE,
+                    "<": _CharFlags.LOWERCASE,
+                    "!": 0,
+                }.get(c, None)
+                if new_flags is not None:
+                    flags = new_flags
+                    continue
+
+                pattern, required_flag = _TEMPLATE_CHARACTERS.get(c, (None, None))
+                if pattern:
+                    char_flags = _CharFlags.REQUIRED if required_flag else _CharFlags(0)
+                    char = self.CharDef(re.compile(pattern), char_flags)
+                else:
+                    char = self.CharDef(
+                        re.compile(re.escape(c)), _CharFlags.SEPARATOR, c
+                    )
+
+            char.flags |= flags
+            self.template.append(char)
+
+        if template_chars:
+            self.blank = template_chars[0]
+
+        if all(char.flags & _CharFlags.SEPARATOR for char in self.template):
+            raise ValueError(
+                "Template must contain at least one non-separator character"
+            )
+
+        self.update_mask(input.placeholder)
+
+    def validate(self, value: str) -> ValidationResult:
+        if self.check(value.ljust(len(self.template), chr(0)), False):
+            return self.success()
+        else:
+            return self.failure("Value does not match template!", value)
+
+    def check(self, value: str, allow_space: bool) -> bool:
+        for c, char_def in zip(value, self.template):
+            if (
+                (char_def.flags & _CharFlags.REQUIRED)
+                and (not char_def.pattern.match(c))
+                and ((c != " ") or not allow_space)
+            ):
+                return False
+        return True
+
+    def insert_separators(self, value: str, cursor_position: int) -> tuple[str, int]:
+        while cursor_position < len(self.template) and (
+            self.template[cursor_position].flags & _CharFlags.SEPARATOR
+        ):
+            value = (
+                value[:cursor_position]
+                + self.template[cursor_position].char
+                + value[cursor_position + 1 :]
+            )
+            cursor_position += 1
+        return value, cursor_position
+
+    def insert_text_at_cursor(self, text: str) -> str | None:
+        value = self.input.value
+        cursor_position = self.input.cursor_position
+        separators = set(
+            [
+                char_def.char
+                for char_def in self.template
+                if char_def.flags & _CharFlags.SEPARATOR
+            ]
+        )
+        for c in text:
+            if c in separators:
+                if c == self.next_separator(cursor_position):
+                    prev_position = self.prev_separator_position(cursor_position)
+                    if (cursor_position > 0) and (prev_position != cursor_position - 1):
+                        next_position = self.next_separator_position(cursor_position)
+                        while cursor_position < next_position + 1:
+                            if (
+                                self.template[cursor_position].flags
+                                & _CharFlags.SEPARATOR
+                            ):
+                                char = self.template[cursor_position].char
+                            else:
+                                char = " "
+                            value = (
+                                value[:cursor_position]
+                                + char
+                                + value[cursor_position + 1 :]
+                            )
+                            cursor_position += 1
+                continue
+            if cursor_position >= len(self.template):
+                break
+            char_def = self.template[cursor_position]
+            assert (char_def.flags & _CharFlags.SEPARATOR) == 0
+            if not char_def.pattern.match(c):
+                return None
+            if char_def.flags & _CharFlags.LOWERCASE:
+                c = c.lower()
+            elif char_def.flags & _CharFlags.UPPERCASE:
+                c = c.upper()
+            value = value[:cursor_position] + c + value[cursor_position + 1 :]
+            cursor_position += 1
+            value, cursor_position = self.insert_separators(value, cursor_position)
+        return value, cursor_position
+
+    def move_cursor(self, delta: int) -> None:
+        cursor_position = self.input.cursor_position
+        if delta < 0 and all(
+            [c.flags & _CharFlags.SEPARATOR for c in self.template[:cursor_position]]
+        ):
+            return
+        cursor_position += delta
+        while (
+            (cursor_position >= 0)
+            and (cursor_position < len(self.template))
+            and (self.template[cursor_position].flags & _CharFlags.SEPARATOR)
+        ):
+            cursor_position += delta
+        self.input.cursor_position = cursor_position
+
+    def delete_at_position(self, position: int | None = None) -> None:
+        value = self.input.value
+        if position is None:
+            position = self.input.cursor_position
+        cursor_position = position
+        if cursor_position < len(self.template):
+            assert (self.template[cursor_position].flags & _CharFlags.SEPARATOR) == 0
+            if cursor_position == len(value) - 1:
+                value = value[:cursor_position]
+            else:
+                value = value[:cursor_position] + " " + value[cursor_position + 1 :]
+        pos = len(value)
+        while pos > 0:
+            char_def = self.template[pos - 1]
+            if ((char_def.flags & _CharFlags.SEPARATOR) == 0) and (
+                value[pos - 1] != " "
+            ):
+                break
+            pos -= 1
+        value = value[:pos]
+        if cursor_position > len(value):
+            cursor_position = len(value)
+        value, cursor_position = self.insert_separators(value, cursor_position)
+        self.input.cursor_position = cursor_position
+        self.input.value = value
+
+    def at_separator(self, position: int | None = None) -> bool:
+        if position is None:
+            position = self.input.cursor_position
+        if (position >= 0) and (position < len(self.template)):
+            return bool(self.template[position].flags & _CharFlags.SEPARATOR)
+        else:
+            return False
+
+    def prev_separator_position(self, position: int | None = None) -> int | None:
+        if position is None:
+            position = self.input.cursor_position
+        for index in range(position - 1, 0, -1):
+            if self.template[index].flags & _CharFlags.SEPARATOR:
+                return index
+        else:
+            return None
+
+    def next_separator_position(self, position: int | None = None) -> int | None:
+        if position is None:
+            position = self.input.cursor_position
+        for index in range(position + 1, len(self.template)):
+            if self.template[index].flags & _CharFlags.SEPARATOR:
+                return index
+        else:
+            return None
+
+    def next_separator(self, position: int | None = None) -> str | None:
+        position = self.next_separator_position(position)
+        if position is None:
+            return None
+        else:
+            return self.template[position].char
+
+    def display(self, value: str) -> str:
+        result = []
+        for c, char_def in zip(value, self.template):
+            if c == " ":
+                c = char_def.char
+            result.append(c)
+        return "".join(result)
+
+    def update_mask(self, placeholder: str) -> None:
+        for index, char_def in enumerate(self.template):
+            if (char_def.flags & _CharFlags.SEPARATOR) == 0:
+                if index < len(placeholder):
+                    char_def.char = placeholder[index]
+                else:
+                    char_def.char = self.blank
+
+    @property
+    def mask(self) -> str:
+        return "".join([c.char for c in self.template])
+
+    @property
+    def empty_mask(self) -> str:
+        return "".join(
+            [
+                " " if (c.flags & _CharFlags.SEPARATOR) == 0 else c.char
+                for c in self.template
+            ]
+        )
 
 
 class Input(Widget, can_focus=True):
@@ -172,7 +463,7 @@ class Input(Widget, can_focus=True):
     """
 
     cursor_blink = reactive(True, init=False)
-    value = reactive("", layout=True, init=False)
+    value: Reactive[str] = reactive("", layout=True, init=False)
     input_scroll_offset = reactive(0)
     cursor_position = reactive(0)
     view_position = reactive(0)
@@ -193,6 +484,8 @@ class Input(Widget, can_focus=True):
     """The maximum length of the input, in characters."""
     valid_empty = var(False)
     """Empty values should pass validation."""
+    template = var("")
+    """Input template currently in use."""
 
     @dataclass
     class Changed(Message):
@@ -253,6 +546,7 @@ class Input(Widget, can_focus=True):
         validators: Validator | Iterable[Validator] | None = None,
         validate_on: Iterable[InputValidationOn] | None = None,
         valid_empty: bool = False,
+        template: str = "",
         name: str | None = None,
         id: str | None = None,
         classes: str | None = None,
@@ -276,6 +570,7 @@ class Input(Widget, can_focus=True):
                 which determine when to do input validation. The default is to do
                 validation for all messages.
             valid_empty: Empty values are valid.
+            template: Optional template string.
             name: Optional name for the input widget.
             id: Optional ID for the widget.
             classes: Optional initial classes for the widget.
@@ -286,6 +581,8 @@ class Input(Widget, can_focus=True):
 
         self._blink_timer: Timer | None = None
         """Timer controlling the blinking of the cursor, instantiated in `on_mount`."""
+
+        self._template: _Template | None = None
 
         self.placeholder = placeholder
         self.highlighter = highlighter
@@ -335,7 +632,13 @@ class Input(Widget, can_focus=True):
             elif self.type == "number":
                 self.validators.append(Number())
 
-        if value is not None:
+        if template:
+            self._template = _Template(self, template)
+        self.template = template
+
+        if (value is not None) or (self._template is not None):
+            if self._template is not None:
+                value, _ = self._template.insert_separators(value or "", 0)
             self.value = value
         if tooltip is not None:
             self.tooltip = tooltip
@@ -365,6 +668,13 @@ class Input(Widget, can_focus=True):
         width = self.content_size.width
         new_view_position = max(0, min(view_position, self.cursor_width - width))
         return new_view_position
+
+    def validate_value(self, value: str) -> str:
+        if self._template is None:
+            return value
+        if not self._template.check(value, True):
+            raise ValueError("Value does not match template!")
+        return value[: len(self._template.mask)]
 
     def _watch_cursor_position(self) -> None:
         width = self.content_size.width
@@ -416,6 +726,18 @@ class Input(Widget, can_focus=True):
         """Repeat validation when valid_empty changes."""
         self._watch_value(self.value)
 
+    def _watch_template(self, template: str) -> None:
+        """Revalidate when template changes."""
+        self._template = _Template(self, template) if template else None
+        if self.is_mounted:
+            self._watch_value(self.value)
+
+    def _watch_placeholder(self, placeholder: str) -> None:
+        """Update template display mask when placeholder changes."""
+        if self._template is not None:
+            self._template.update_mask(placeholder)
+            self.refresh()
+
     def validate(self, value: str) -> ValidationResult | None:
         """Run all the validators associated with this Input on the supplied value.
 
@@ -437,8 +759,12 @@ class Input(Widget, can_focus=True):
             self.set_class(not valid, "-invalid")
             self.set_class(valid, "-valid")
 
+        validators = list(self.validators)
+        if self._template is not None:
+            validators.append(self._template)
+
         # If no validators are supplied, and therefore no validation occurs, we return None.
-        if not self.validators:
+        if not validators:
             self._valid = True
             set_classes()
             return None
@@ -449,7 +775,7 @@ class Input(Widget, can_focus=True):
             return None
 
         validation_results: list[ValidationResult] = [
-            validator.validate(value) for validator in self.validators
+            validator.validate(value) for validator in validators
         ]
         combined_result = ValidationResult.merge(validation_results)
         self._valid = combined_result.is_valid
@@ -471,7 +797,7 @@ class Input(Widget, can_focus=True):
 
     def render(self) -> RenderResult:
         self.view_position = self.view_position
-        if not self.value:
+        if (not self.value) and (self._template is None):
             placeholder = Text(self.placeholder, justify="left")
             placeholder.stylize(self.get_component_rich_style("input--placeholder"))
             if self.has_focus:
@@ -491,7 +817,11 @@ class Input(Widget, can_focus=True):
         if self.password:
             return Text("•" * len(self.value), no_wrap=True, overflow="ignore")
         else:
-            text = Text(self.value, no_wrap=True, overflow="ignore")
+            if self._template is None:
+                value = self.value
+            else:
+                value = self._template.display(self.value)
+            text = Text(value, no_wrap=True, overflow="ignore")
             if self.highlighter is not None:
                 text = self.highlighter(text)
             return text
@@ -558,6 +888,8 @@ class Input(Widget, can_focus=True):
             cell_offset += cell_width
         else:
             self.cursor_position = len(self.value)
+        if (self._template is not None) and self._template.at_separator():
+            self._template.move_cursor(1)
 
     async def _on_suggestion_ready(self, event: SuggestionReady) -> None:
         """Handle suggestion messages and set the suggestion when relevant."""
@@ -590,7 +922,18 @@ class Input(Widget, can_focus=True):
             # Character is allowed
             return True
 
-        if self.cursor_position >= len(self.value):
+        if self._template is not None:
+            new_value = self._template.insert_text_at_cursor(text)
+            if new_value is not None:
+                new_value, cursor_position = new_value
+                if check_allowed_value(new_value):
+                    self.value = new_value
+                    self.cursor_position = cursor_position
+                else:
+                    self.restricted()
+            else:
+                self.restricted()
+        elif self.cursor_position >= len(self.value):
             new_value = self.value + text
             if check_allowed_value(new_value):
                 self.value = new_value
@@ -618,11 +961,19 @@ class Input(Widget, can_focus=True):
 
     def clear(self) -> None:
         """Clear the input."""
-        self.value = ""
+        if self._template is None:
+            value, cursor_position = "", 0
+        else:
+            value, cursor_position = self._template.insert_separators("", 0)
+        self.value = value
+        self.cursor_position = cursor_position
 
     def action_cursor_left(self) -> None:
         """Move the cursor one position to the left."""
-        self.cursor_position -= 1
+        if self._template is None:
+            self.cursor_position -= 1
+        else:
+            self._template.move_cursor(-1)
 
     def action_cursor_right(self) -> None:
         """Accept an auto-completion or move the cursor one position to the right."""
@@ -630,11 +981,17 @@ class Input(Widget, can_focus=True):
             self.value = self._suggestion
             self.cursor_position = len(self.value)
         else:
-            self.cursor_position += 1
+            if self._template is None:
+                self.cursor_position += 1
+            else:
+                self._template.move_cursor(1)
 
     def action_home(self) -> None:
         """Move the cursor to the start of the input."""
-        self.cursor_position = 0
+        if self._template is None:
+            self.cursor_position = 0
+        else:
+            self._template.move_cursor(-len(self.template))
 
     def action_end(self) -> None:
         """Move the cursor to the end of the input."""
@@ -644,7 +1001,17 @@ class Input(Widget, can_focus=True):
 
     def action_cursor_left_word(self) -> None:
         """Move the cursor left to the start of a word."""
-        if self.password:
+        if self._template is not None:
+            if self._template.at_separator(self.cursor_position - 1):
+                position = self._template.prev_separator_position(
+                    self.cursor_position - 1
+                )
+            else:
+                position = self._template.prev_separator_position()
+            if position:
+                position += 1
+            self.cursor_position = position or 0
+        elif self.password:
             # This is a password field so don't give any hints about word
             # boundaries, even during movement.
             self.action_home()
@@ -660,7 +1027,13 @@ class Input(Widget, can_focus=True):
 
     def action_cursor_right_word(self) -> None:
         """Move the cursor right to the start of a word."""
-        if self.password:
+        if self._template is not None:
+            position = self._template.next_separator_position()
+            if position is None:
+                self.cursor_position = len(self._template.mask)
+            else:
+                self.cursor_position = position + 1
+        elif self.password:
             # This is a password field so don't give any hints about word
             # boundaries, even during movement.
             self.action_end()
@@ -673,16 +1046,29 @@ class Input(Widget, can_focus=True):
 
     def action_delete_right(self) -> None:
         """Delete one character at the current cursor position."""
-        value = self.value
-        delete_position = self.cursor_position
-        before = value[:delete_position]
-        after = value[delete_position + 1 :]
-        self.value = f"{before}{after}"
-        self.cursor_position = delete_position
+        if self._template is not None:
+            self._template.delete_at_position()
+        else:
+            value = self.value
+            delete_position = self.cursor_position
+            before = value[:delete_position]
+            after = value[delete_position + 1 :]
+            self.value = f"{before}{after}"
+            self.cursor_position = delete_position
 
     def action_delete_right_word(self) -> None:
         """Delete the current character and all rightward to the start of the next word."""
-        if self.password:
+        if self._template:
+            position = self._template.next_separator_position()
+            if position is not None:
+                position += 1
+            else:
+                position = len(self.value)
+            for index in range(self.cursor_position, position):
+                self.cursor_position = index
+                if not self._template.at_separator():
+                    self._template.delete_at_position()
+        elif self.password:
             # This is a password field so don't give any hints about word
             # boundaries, even during deletion.
             self.action_delete_right_all()
@@ -705,7 +1091,10 @@ class Input(Widget, can_focus=True):
         if self.cursor_position <= 0:
             # Cursor at the start, so nothing to delete
             return
-        if self.cursor_position == len(self.value):
+        if self._template is not None:
+            self._template.move_cursor(-1)
+            self._template.delete_at_position()
+        elif self.cursor_position == len(self.value):
             # Delete from end
             self.value = self.value[:-1]
             self.cursor_position = len(self.value)
@@ -722,7 +1111,23 @@ class Input(Widget, can_focus=True):
         """Delete leftward of the cursor position to the start of a word."""
         if self.cursor_position <= 0:
             return
-        if self.password:
+        if self._template is not None:
+            if self._template.at_separator(self.cursor_position - 1):
+                position = self._template.prev_separator_position(
+                    self.cursor_position - 1
+                )
+            else:
+                position = self._template.prev_separator_position()
+            if position:
+                position += 1
+            else:
+                position = 0
+            for index in range(position, self.cursor_position):
+                self.cursor_position = index
+                if not self._template.at_separator():
+                    self._template.delete_at_position()
+            self.cursor_position = position
+        elif self.password:
             # This is a password field so don't give any hints about word
             # boundaries, even during deletion.
             self.action_delete_left_all()
@@ -741,7 +1146,17 @@ class Input(Widget, can_focus=True):
     def action_delete_left_all(self) -> None:
         """Delete all characters to the left of the cursor position."""
         if self.cursor_position > 0:
-            self.value = self.value[self.cursor_position :]
+            if self._template is not None:
+                cursor_position = self.cursor_position
+                if cursor_position >= len(self.value):
+                    self.value = ""
+                else:
+                    self.value = (
+                        self._template.empty_mask[:cursor_position]
+                        + self.value[cursor_position:]
+                    )
+            else:
+                self.value = self.value[self.cursor_position :]
             self.cursor_position = 0
 
     async def action_submit(self) -> None: